Update ModronGUI.ahk
The F-Key array gets filled before reading the data, causing the array to end up empty. As a result, the GUI falls back to default values when rendering (everything's ticked) and saving settings, though it still works when running because then a different variable gets used which has the expected values assigned.